+++ examples/sample-container/src/main/kotlin/de/gtarc/opaca/sample/Main.kt
@@ -1,9 +1,9 @@
 package de.gtarc.opaca.sample
 
-import de.gtarc.opaca.container.ContainerAgent
-import de.gtarc.opaca.util.ConfigLoader
 import de.dailab.jiacvi.communication.LocalBroker
 import de.dailab.jiacvi.dsl.agentSystem
+import de.gtarc.opaca.container.ContainerAgent
+import de.gtarc.opaca.util.ConfigLoader
 
 fun main() {
     val image = ConfigLoader.loadContainerImageFromResources("/sample-image.json")