+++ examples/sample-container/src/main/kotlin/de/gtarc/opaca/sample/SampleAgent.kt
@@ -1,17 +1,12 @@
 package de.gtarc.opaca.sample
 
+import de.dailab.jiacvi.behaviour.act
 import de.gtarc.opaca.api.AgentContainerApi
 import de.gtarc.opaca.container.AbstractContainerizedAgent
-import de.gtarc.opaca.model.Stream
 import de.gtarc.opaca.model.Message
-import de.dailab.jiacvi.behaviour.act
-
+import de.gtarc.opaca.model.Parameter
 import java.io.ByteArrayInputStream
-import java.io.ByteArrayOutputStream
-import java.io.InputStream
 import java.nio.charset.Charset
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import de.gtarc.opaca.container.OpacaServer.OpacaException
 import java.io.IOException
 
@@ -24,34 +19,59 @@ class SampleAgent(name: String): AbstractContainerizedAgent(name=name) {
     override fun preStart() {
         super.preStart()
 
-        addAction("DoThis", mapOf("message" to "String", "sleep_seconds" to "Int"), "String") {
+        addAction("DoThis", mapOf(
+            "message" to Parameter("string", true),
+            "sleep_seconds" to Parameter("integer", true)
+        ), "string") {
             actionDoThis(it["message"]!!.asText(), it["sleep_seconds"]!!.asInt())
         }
-        addAction("GetInfo", mapOf(), "Map") {
+        addAction("GetInfo", mapOf(), "object") {
             actionGetInfo()
         }
-        addAction("GetEnv", mapOf(), "Map") {
+        addAction("GetEnv", mapOf(), "object") {
             actionGetEnv()
         }
-        addAction("Add", mapOf("x" to "String", "y" to "Int"), "Int") {
+        addAction("Add", mapOf(
+            "x" to Parameter("integer", true),
+            "y" to Parameter("integer", true)
+        ), "integer") {
             actionAdd(it["x"]!!.asInt(), it["y"]!!.asInt())
         }
-        addAction("Fail", mapOf(), "void") {
+        addAction("Fail", mapOf(), "null") {
             actionFail()
         }
-        addAction("CreateAction", mapOf("name" to "String", "notify" to "Boolean"), "void") {
-            createAction(it["name"]!!.asText(), it["notify"]!!.asBoolean())
+        addAction("CreateAction", mapOf(
+            "name" to Parameter("string", true),
+            "notify" to Parameter("boolean", false)
+        ), "null") {
+            createAction(it["name"]!!.asText(), it["notify"]?.asBoolean() ?: true)
         }
-        addAction("SpawnAgent", mapOf("name" to "String"), "void") {
+        addAction("SpawnAgent", mapOf(
+            "name" to Parameter("string", true)
+        ), "null") {
             spawnAgent(it["name"]!!.asText())
         }
-        addAction("Deregister", mapOf(), "void") {
+        addAction("Deregister", mapOf(), "null") {
             deregister(false)
             stop()
         }
-        addAction("ErrorTest", mapOf("hint" to "String"), "String") {
-            val hint = it["hint"]!!.asText()
-            actionErrorTest(hint)
+        addAction("ErrorTest", mapOf(
+            "hint" to Parameter("string", true)
+        ), "string") {
+            actionErrorTest(it["hint"]!!.asText())
+        }
+        addAction("ValidatorTest", mapOf(
+            "car" to Parameter("Car", true),
+            "listOfLists" to Parameter("array", true,
+                Parameter.ArrayItems("array", Parameter.ArrayItems("integer", null))),
+            "decimal" to Parameter("number", false),
+            "desk" to Parameter("Desk", false)
+        ), "string") {
+            val carText = "Parameter \"car\": ${it["car"]!!.asText()}"
+            val listText = "Parameter \"listOfLists\"${it["listOfLists"]!!.asText()}"
+            val result = "ValidatorTest:\n$carText\n$listText"
+            print(result)
+            result
         }
 
         addStreamPost("PostStream", this::actionPostStream)
@@ -82,12 +102,12 @@ class SampleAgent(name: String): AbstractContainerizedAgent(name=name) {
         lastPostedStream = content
     }
 
-    private fun actionDoThis(message: String, sleep_seconds: Int): String {
+    private fun actionDoThis(message: String, sleepSeconds: Int): String {
         log.info("in 'DoThis' action, waiting...")
         println(message)
-        Thread.sleep(1000 * sleep_seconds.toLong())
+        Thread.sleep(1000 * sleepSeconds.toLong())
         log.info("done waiting")
-        return "Action 'DoThis' of $name called with message=$message and sleep_seconds=$sleep_seconds"
+        return "Action 'DoThis' of $name called with message=$message and sleep_seconds=$sleepSeconds"
     }
 
     private fun actionAdd(x: Int, y: Int) = x + y
@@ -110,7 +130,7 @@ class SampleAgent(name: String): AbstractContainerizedAgent(name=name) {
     private fun actionGetEnv() = System.getenv()
 
     private fun createAction(name: String, notify: Boolean) {
-        addAction(name, mapOf(), "String") {
+        addAction(name, mapOf(), "string") {
             "Called extra action $name"
         }
         register(notify)