+++ opaca-platform/src/main/java/de/gtarc/opaca/platform/PlatformImpl.java
@@ -66,6 +66,9 @@ public class PlatformImpl implements RuntimePlatformApi {
     /** Set of remote Runtime Platform URLs with a pending connection request */
     private final Set<String> pendingConnections = new HashSet<>();
 
+    /** Map of validators for validating action argument types for each container */
+    private final Map<String, ArgumentValidator> validators = new HashMap<>();
+
 
     @PostConstruct
     public void initialize() {
@@ -90,6 +93,11 @@ public class PlatformImpl implements RuntimePlatformApi {
         this.containerClient.initialize(config, sessionData);
         this.containerClient.testConnectivity();
         // TODO add list of known used ports to config (e.g. the port of the RP itself, or others)
+
+        for (var containerId : runningContainers.keySet()) {
+            var image = runningContainers.get(containerId).getImage();
+            validators.put(containerId, new ArgumentValidator(image));
+        }
     }
 
     @Override
@@ -135,7 +143,7 @@ public class PlatformImpl implements RuntimePlatformApi {
 
     @Override
     public void send(String agentId, Message message, String containerId, boolean forward) throws IOException, NoSuchElementException {
-        var clients = getClients(containerId, agentId, null, null, forward);
+        var clients = getClients(containerId, agentId, null, null, null, forward);
 
         IOException lastException = null;
         for (ApiProxy client: (Iterable<? extends ApiProxy>) clients::iterator) {
@@ -154,7 +162,7 @@ public class PlatformImpl implements RuntimePlatformApi {
 
     @Override
     public void broadcast(String channel, Message message, String containerId, boolean forward) {
-        var clients = getClients(containerId, null, null, null, forward);
+        var clients = getClients(containerId, null, null, null, null, forward);
 
         for (ApiProxy client: (Iterable<? extends ApiProxy>) clients::iterator) {
             log.info("Forwarding /broadcast to " + client.baseUrl);
@@ -173,7 +181,7 @@ public class PlatformImpl implements RuntimePlatformApi {
 
     @Override
     public JsonNode invoke(String action, Map<String, JsonNode> parameters, String agentId, int timeout, String containerId, boolean forward) throws IOException, NoSuchElementException {
-        var clients = getClients(containerId, agentId, action, null, forward);
+        var clients = getClients(containerId, agentId, action, parameters, null, forward);
 
         IOException lastException = null;
         for (ApiProxy client: (Iterable<? extends ApiProxy>) clients::iterator) {
@@ -188,7 +196,7 @@ public class PlatformImpl implements RuntimePlatformApi {
             }
         }
         if (lastException != null) throw lastException;
-        throw new NoSuchElementException(String.format("Not found: action '%s' @ agent '%s'", action, agentId));
+        throw new NoSuchElementException(String.format("Not found: action '%s' @ agent '%s', or the given parameters are invalid.", action, agentId));
     }
 
     @Override
@@ -198,7 +206,7 @@ public class PlatformImpl implements RuntimePlatformApi {
 
     @Override
     public ResponseEntity<StreamingResponseBody> getStream(String stream, String agentId, String containerId, boolean forward) throws IOException {
-        var clients = getClients(containerId, agentId, null, stream, forward);
+        var clients = getClients(containerId, agentId, null, null, stream, forward);
         
         IOException lastException = null;
         for (ApiProxy client: (Iterable<? extends ApiProxy>) clients::iterator) {
@@ -227,7 +235,7 @@ public class PlatformImpl implements RuntimePlatformApi {
 
     @Override
     public void postStream(String stream, byte[] inputStream, String agentId, String containerId, boolean forward) throws IOException {
-        var clients = getClients(containerId, agentId, null, stream, forward);
+        var clients = getClients(containerId, agentId, null, null, stream, forward);
         
         IOException lastException = null;
         for (ApiProxy client: (Iterable<? extends ApiProxy>) clients::iterator) {
@@ -273,6 +281,7 @@ public class PlatformImpl implements RuntimePlatformApi {
                 runningContainers.put(agentContainerId, container);
                 startedContainers.put(agentContainerId, postContainer);
                 tokens.put(agentContainerId, token);
+                validators.put(agentContainerId, new ArgumentValidator(container.getImage()));
                 container.setOwner(owner);
                 userDetailsService.createUser(agentContainerId, agentContainerId,
                         config.enableAuth ? userDetailsService.getUserRole(owner) : Role.GUEST,
@@ -335,6 +344,7 @@ public class PlatformImpl implements RuntimePlatformApi {
         if (container == null) return false;
         runningContainers.remove(containerId);
         startedContainers.remove(containerId);
+        validators.remove(containerId);
         userDetailsService.removeUser(containerId);
         containerClient.stopContainer(containerId);
         notifyConnectedPlatforms();
@@ -415,6 +425,7 @@ public class PlatformImpl implements RuntimePlatformApi {
             var containerInfo = client.getContainerInfo();
             containerInfo.setConnectivity(runningContainers.get(containerId).getConnectivity());
             runningContainers.put(containerId, containerInfo);
+            validators.put(containerId, new ArgumentValidator(containerInfo.getImage()));
             notifyConnectedPlatforms();
             return true;
         } catch (IOException e) {
@@ -482,29 +493,33 @@ public class PlatformImpl implements RuntimePlatformApi {
      * @param includeConnected Whether to also forward to connected Runtime Platforms
      * @return list of clients to send requests to these valid containers/platforms
      */
-    private Stream<ApiProxy> getClients(String containerId, String agentId, String action, String stream, boolean includeConnected) {
+    private Stream<ApiProxy> getClients(String containerId, String agentId, String action, Map<String, JsonNode> parameters, String stream, boolean includeConnected) {
         // local containers
         var containerClients = runningContainers.values().stream()
-                .filter(c -> matches(c, containerId, agentId, action, stream))
+                .filter(c -> matches(c, containerId, agentId, action, parameters, stream))
                 .map(c -> getClient(c.getContainerId(), tokens.get(c.getContainerId())));
 
         if (!includeConnected) return containerClients;
 
         // remote platforms
         var platformClients = connectedPlatforms.entrySet().stream()
-            .filter(entry -> entry.getValue().getContainers().stream().anyMatch(c -> matches(c, containerId, agentId, action, stream)))
-            .map(entry -> {return new ApiProxy(entry.getKey(), tokens.get(entry.getKey()));});
+            .filter(entry -> entry.getValue().getContainers().stream().anyMatch(c -> matches(c, containerId, agentId, action, parameters, stream)))
+            .map(entry -> new ApiProxy(entry.getKey(), tokens.get(entry.getKey())));
 
         return Stream.concat(containerClients, platformClients);
     }
     /**
      * Check if Container ID matches and has matching agent and/or action.
      */
-    private boolean matches(AgentContainer container, String containerId, String agentId, String action, String stream) {
+    private boolean matches(AgentContainer container, String containerId, String agentId, String action, Map<String, JsonNode> arguments, String stream) {
+        var validator = validators.containsKey(container.getContainerId())
+                ? validators.get(container.getContainerId()) // own container
+                : new ArgumentValidator(container.getImage()); // connected rp container
         return (containerId == null || container.getContainerId().equals(containerId)) &&
                 container.getAgents().stream()
                         .anyMatch(a -> (agentId == null || a.getAgentId().equals(agentId))
-                                && (action == null || a.getActions().stream().anyMatch(x -> x.getName().equals(action)))
+                                && (action == null || a.getActions().stream().anyMatch(x -> x.getName().equals(action)
+                                    && (arguments == null || (validator != null && validator.isArgsValid(x.getParameters(), arguments)))))
                                 && (stream == null || a.getStreams().stream().anyMatch(x -> x.getName().equals(stream)))
                         );
     }